refactor(theme): rename /theme dir to /styles
- mirrors Mui
