+++ libs/spark/src/InputLabel.ts
@@ -1,4 +1,4 @@
-import { palette } from './theme/palette';
+import { palette } from './styles/palette';
 export const MuiInputLabelStyleOverrides = {
   root: {
     color: palette.text.onLight,
