+++ libs/spark/src/SparkThemeProvider.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { ThemeProvider, StylesProvider } from '@material-ui/core/styles';
 import { ThemeProvider as StyledThemeProvider } from 'styled-components';
-import { theme } from './theme';
+import { theme } from './styles';
 import { CssBaseline } from '@material-ui/core';
 
 export const SparkThemeProvider = (props) => {
