+++ libs/spark/src/styles/palette.ts
