+++ libs/spark/src/styles/props.ts
